Added tests for new reducer
