+++ test/unit/src/common/reducers/entities/t_snap.js
@@ -2,68 +2,12 @@ import expect from 'expect';
 
 import * as RegisterNameActionTypes from '../../../../../../src/common/actions/register-name';
 import snap from '../../../../../../src/common/reducers/entities/snap.js';
-import * as ActionTypes from '../../../../../../src/common/actions/register-name';
 
 describe('snaps entities', function() {
 
   let state = {
     storeName: 'test-name'
   };
-
-  context('name ownership actions', () => {
-    let state = {
-      snapcraftData: {
-        name: 'test-name'
-      }
-    };
-
-    const requestAction = {
-      type: ActionTypes.CHECK_NAME_OWNERSHIP_REQUEST,
-      payload: {
-        id: 'snapId',
-        snapName: 'test-name',
-      }
-    };
-
-    const successAction = {
-      type: ActionTypes.CHECK_NAME_OWNERSHIP_SUCCESS,
-      payload: {
-        id: 'snapId',
-        snapName: 'test-name',
-        status: 'test-status'
-      }
-    };
-
-    const errorAction = {
-      type: ActionTypes.CHECK_NAME_OWNERSHIP_ERROR,
-      payload: {
-        id: 'snapId',
-        snapName: 'test-name',
-        status: 'test-status'
-      }
-    };
-
-    it('CHECK_NAME_OWNERSHIP_REQUEST should update isFetching state', function() {
-      expect(snap(state, requestAction).snapcraftData.isFetching).toBe(true);
-    });
-
-    it('CHECK_NAME_OWNERSHIP_SUCCESS should update isFetching state', function() {
-      expect(snap(state, successAction).snapcraftData.isFetching).toBe(false);
-    });
-
-    it('CHECK_NAME_OWNERSHIP_SUCCESS should set name ownership status', function() {
-      expect(snap(state, successAction).snapcraftData.nameOwnershipStatus).toBe('test-status');
-    });
-
-    it('CHECK_NAME_OWNERSHIP_ERROR should update isFetching state', function() {
-      expect(snap(state, errorAction).snapcraftData.isFetching).toBe(false);
-    });
-
-    it('CHECK_NAME_OWNERSHIP_ERROR should reset name ownership status', function() {
-      expect(snap(state, errorAction).snapcraftData.nameOwnershipStatus).toBe(null);
-    });
-  
-  });
 
   context('on register name actions', () => {
     const initialStatus = {
